distinguishes utils and constants, updates imports
